front: intervals editor, create separated components for the scales and the interval items
